requirements.txt
openai
python-dotenv
gradio==3.50.2
llama-index==0.10.20
chromadb
jinja2==3.1.3
llama-index-vector-stores-chroma
